fix: preserve the refusal counter dtype across cell mapping
Summing mapped int32 counters otherwise promotes them to the default integer width in double precision. Pin the mesh-level reduction to int32 and assert the public counter dtype in the exercised overflow-refusal test. Hold the existing allocation for final-source validation.

+++ nova/equilibrium/separatrix_clip.py
@@ -1003,7 +1003,7 @@ def _traced_clip(
         contour_area=jnp.abs(jnp.sum(result.contour_area)),
         patch_area_sum=jnp.sum(result.area),
         vertex_capacity=result.vertex_capacity[0],
-        refused_cell_count=jnp.sum(result.refused_cell_count),
+        refused_cell_count=jnp.sum(result.refused_cell_count, dtype=jnp.int32),
     )
